Fix: window undefined 에러 임시 해결
isBrowser 변수로 if문처리

+++ src/utils/browserStorage.ts
@@ -1,3 +1,5 @@
+import { isBrowser } from '@/utils'
+
 export interface BrowserStorage {
   set: <T extends string>(key: string, value: T) => void
   get: (key: string) => string | null
@@ -7,16 +9,16 @@ export interface BrowserStorage {
 const browserStorage: BrowserStorage = {
   set: (key, value) => {
     if (typeof value === 'string') {
-      localStorage.setItem(key, value)
+      isBrowser && localStorage.setItem(key, value)
     } else {
-      localStorage.setItem(key, JSON.stringify(value))
+      isBrowser && localStorage.setItem(key, JSON.stringify(value))
     }
   },
   get: key => {
-    return localStorage.getItem(key) || null
+    return (isBrowser && localStorage.getItem(key)) || null
   },
   reset: () => {
-    localStorage.clear()
+    isBrowser && localStorage.clear()
   },
 }
